resolve review

+++ src/Native/src/kernels/stackvm/optimized/slice.cpp
@@ -49,7 +49,7 @@ void _slice_contiguous_dim_copy(const axes_t &begins,
 }
 
 // optimized for n c h 1
-unsigned long inline squeeze_dims(const gsl::span<const size_t> &in_shape) {
+size_t inline squeeze_dims(const gsl::span<const size_t> &in_shape) {
     return in_shape[in_shape.size() - 1] == 1 ? in_shape.size() - 1 - 1
                                               : in_shape.size() - 1;
 }
